00_notebook_viewer.v4/requirements.txt
# Research Platform - Enterprise Edition
# Requirements

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# API & Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Templates & Static Files
jinja2>=3.1.3
aiofiles>=23.2.1

# Markdown Processing
markdown>=3.5.2
pymdown-extensions>=10.7
bleach>=6.1.0

# PDF Generation
playwright>=1.41.0

# Utilities
python-dateutil>=2.8.2
httpx>=0.26.0

# Development (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Production
gunicorn>=21.2.0
